# src/swagger/swagger.yaml
definitions:
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse:
    properties:
      error:
        example: error description
        type: string
    type: object
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.IDResponse:
    properties:
      id:
        type: integer
    type: object
...
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterKPRequest:
    properties:
      kp_id:
        type: string
    type: object
...
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterRequest:
    properties:
      chrono:
        type: integer
      genres:
        items:
          type: string
        type: array
      image_url:
        type: string
      kp_id:
        type: string
      name:
        type: string
      year:
        type: integer
    type: object
  git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterResponse:
    properties:
      chrono:
        type: integer
      createdat:
        description: will not be used, satisfy musttag linter
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: integer
      image_url:
        type: string
      kp_id:
        type: string
      name:
        type: string
      userId:
        type: integer
      year:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/v2/posters:
    post:
      consumes:
      - application/json
      description: create poster
      parameters:
      - description: Poster body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterRequest'
      - description: TG-ID token
        in: header
        name: X-User-Token
        required: true
        type: string
      responses:
        "201":
          description: id
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.IDResponse'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
      summary: Create
      tags:
      - posters/v2
  /api/v2/posters/{id}:
    delete:
      consumes:
      - application/json
      description: delete poster
      parameters:
      - description: TG-ID token
        in: header
        name: X-User-Token
        required: true
        type: string
      - description: PosterId
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
        "400":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
      summary: Delete
      tags:
      - posters/v2
    get:
      consumes:
      - application/json
      description: get poster
      parameters:
      - description: TG-ID token
        in: header
        name: X-User-Token
        required: true
        type: string
      - description: PosterId
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Poster
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterResponse'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
      summary: Get
      tags:
      - posters/v2
    put:
      consumes:
      - application/json
      description: update poster
      parameters:
      - description: Poster body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterRequest'
      - description: PosterId
        in: path
        name: id
        required: true
        type: integer
      - description: TG-ID token
        in: header
        name: X-User-Token
        required: true
        type: string
      responses:
        "200":
          description: OK
        "400":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "404":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
      summary: Update
      tags:
      - posters/v2
  /api/v2/posters/kp:
    post:
      consumes:
      - application/json
      description: create poster
      parameters:
      - description: KP ID body
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.PosterKPRequest'
      - description: TG-ID token
        in: header
        name: X-User-Token
        required: true
        type: string
      responses:
        "201":
          description: id
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.IDResponse'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "401":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
        "500":
          description: Error
          schema:
            $ref: '#/definitions/git_iu7_bmstu_ru_vai20u117_testing_src_internal_api_v2_model.ErrorResponse'
      summary: Create poster based on KP ID
      tags:
      - posters/v2
...
swagger: "2.0"